+++ applications/frontend/app/favicon.ico
