+++ applications/frontend/app/globals.css
