+++ applications/frontend/app/layout.tsx
