+++ applications/frontend/assets/background-planet.png
