+++ applications/frontend/assets/background-surface.png
